Fixed some issues and continuing to finalize for testing

+++ LL_Project/Lichs Lair/Assets/Code/Player/PlayerData.cs
@@ -6,7 +6,7 @@ using UnityEngine;
 public class PlayerData : ScriptableObject
 {
     
-    public int CurrentHealth;
+    public float CurrentHealth;
     public int CurrentMana;
